pyproject.toml
[tool.poetry]
name = "ine5420"
version = "0.1.0"
description = ""
authors = ["caiopo <caiopoliveira@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pycairo = "^1.18"
pygobject = "^3.32"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
